Move proxy endpoints to a separate controller
